Moved Conversion-classes into seperate files
